+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
 
+- Add support for rules, backendRefs and matches.
+- Initialize project
 
 [Unreleased]: https://github.com/giantswarm/ga-migration-cli/tree/main